fix(client): Google 매트릭스 duration 파싱·빈 목적지 방어 보강

+++ src/main/java/com/dnd/moyeolak/global/client/google/dto/RouteMatrixEntry.java
@@ -19,7 +19,7 @@ public record RouteMatrixEntry(
     }
 
     public long durationSeconds() {
-        return Long.parseLong(duration.replace("s", ""));
+        return (long) Double.parseDouble(duration.replace("s", ""));
     }
 
     public int safeDistanceMeters() {
